Engine: rewrote custom PACKFILE without use of normal member
This complements a1329085d94e3723e4cbb960a8b1818c1d9a00dc

I made a mistake by keep using normal.todo member of PACKFILE, it appears you cannot do this if you create one with a custom vtable. Also, custom pack functions were not checking for remaining bytes in the asset segment.
This commit fixes these issues by adding more fields to PACKFILE's userdata and making custom pack functions rely on these.

+++ Engine/media/audio/clip_mymp3.h
@@ -24,7 +24,7 @@ struct MYMP3:public SOUNDCLIP
 {
     ALMP3_MP3STREAM *stream;
     PACKFILE *in;
-    long  filesize;
+    size_t filesize;
     char *buffer;
     int chunksize;
